remove capabilities on create and update tests
